pyproject.toml
[project]
name = "jwtauth"
version = "1.0.0"
description = """
this is the package for the fastapi app to support jwt authentication.
"""
authors = [
    {name = "Narayan",email = "narayanh7888@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyjwt (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "fastapi (>=0.115.7,<0.116.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "cryptography (>=44.0.0,<45.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "python-jose (>=3.3.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
